+++ ts/packages/core/src/utils/transformers/connectedAccounts.ts
@@ -7,6 +7,7 @@ import {
   ConnectedAccountRetrieveResponseSchema,
 } from '../../types/connectedAccounts.types';
 import { ValidationError } from '../../errors';
+import logger from '../logger';
 
 /**
  * Transforms the raw connected account response from the Composio API to the SDK format.
@@ -23,7 +24,7 @@ import { ValidationError } from '../../errors';
 export function transformConnectedAccountResponse(
   response: RawConnectedAccountRetrieveResponse | RawConnectedAccountListResponse['items'][0]
 ): ConnectedAccountRetrieveResponse {
-  const result = ConnectedAccountRetrieveResponseSchema.safeParse({
+  const responseToParse = {
     ...response,
     authConfig: {
       ...response.auth_config,
@@ -38,11 +39,15 @@ export function transformConnectedAccountResponse(
     createdAt: response.created_at,
     updatedAt: response.updated_at,
     testRequestEndpoint: response.test_request_endpoint,
-  });
+  };
+  const result = ConnectedAccountRetrieveResponseSchema.safeParse(responseToParse);
   if (!result.success) {
-    throw new ValidationError('Failed to parse connected account retrieve response', {
+    logger.error('Failed to parse connected account retrieve response', {
       cause: result.error,
     });
+    // this is a fallback for when the response is not valid, it will return the response as is
+    // ideally to not break the app if there are some issues with the response, fields missing etc
+    return responseToParse as unknown as ConnectedAccountRetrieveResponse;
   }
   return result.data;
 }